1) Removido classes de testes duplicadas unificando as mesmas 2) Alterado nome da classe de teste de CPFTest para PhoneTest, pois o nome do arquivo é PhoneTest.php
